pyproject.toml
[tool.poetry]
name = "telegrify"
version = "0.9.0"
description = "Simple Telegram notification framework with plugin support"
authors = ["Gemechis Chala <gladsonchala@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/venopyx/telegrify"
repository = "https://github.com/venopyx/telegrify"
documentation = "https://github.com/venopyx/telegrify#readme"
keywords = ["telegram", "notifications", "bot", "webhook", "fastapi", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
packages = [{include = "telegrify"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pyyaml = "^6.0"
jinja2 = "^3.1.0"
python-dotenv = "^1.0.0"
click = "^8.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[tool.poetry.scripts]
telegrify = "telegrify.cli.commands:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
